observability/observability-assistant/requirements.txt
streamlit>=1.32.0
boto3>=1.28.0
watchdog>=3.0.0 
strands-agents>=0.1.9
strands-agents-tools
pydantic>=2.0.0
mcp>=1.12.0
python-dotenv>=1.0.0
